The mapServer now caches the mapped quad relations both in memory and in files, for which it creates a cache directory (per graph / config file).
It only keeps relations in memory for a limited number of quads, as specified in the config file in the memoryMaxQuadsToCache parameter.

In addition it offers monitoring and warnings for memory usage.

The memory stats can be retrieved with an API:
http://host:port/baseUri/stats
http://host:port/baseUri/stats?view=complete (with quad details)

Also an alert monitor page can be used for use in service watchers like uptime robot or pingdom, this page is available at:
http://host:port/baseUri/memoryMonitor

For use with watchers this page only returns a 200 response with the text WARNING when the mapServer node process exceeds the memory usage specified in the config file in parameter memoryUsageWarningThreshold.