Changes for markers and navigation rectangle
When marker offsets are not sufficient, we introduce the concept of defining covering area's floating on top of the graph.
Markers will stick to the edges of these covering area's when the corresponding node text is covered by such an area.

The navigation / selection rectangle now consists of actually 2 rectangles of contrasting colors, so the rectangle will have good visibility regardless of the underlying graph colors.

Also solved a minor bug, the original nodeId when not supplied in a option (but URL par) needed to be stored in the options too.