Allow for zoom beyond max zoom level, fix node names vanishing in large zoom, much improved collision detection for text boxes, implement collision check for node names not in current selected cluster, option to use non scaling strokes to prevent using CSS or JS at all for reverse scaling nodes, added option to only show a bitmap.
Detailed description
--------------------
Improved collision detection for text boxes, plus detection of collision of 'marker' text and node texts.

Added an option to only show a bitmap.

Added an option to zoom beyond the max zoom level (graph minscale) :

Nodes and text are scaled back when zooming, scaling the graph svg.

Nodes scaled with CSS for performance (lots of nodes), based on px and em’s settings per zoom level.
Node names are scaled with JS as these are only displayed for highlighted nodes.

Works on WebKit (Chrome and Safari). IE and FF have limits for a minimum PX size, otherwise the nodes and text will disappear.
For this now additional back scaling using transform is implemented beyond the min PX size.

This has complication that scaling is done relative to the elements origin, which is different for every element.
Basically origin correction is the (x, y) position which is now used in a matrix transform to keep nodes and texts in place while scaling.

This is implemented using JS because the origin is element bound. This reverse scaling is done only for nodes in the viewport, using a safety (half of the avg quad width) to scale preemptive for expected panning movements.

This still has some issues which need to be solved:
- IE: nodes are sometimes not reverse scaled while in viewport
- FF: clicks on nodes not working on deepest zoom levels

For IE also the edge-width had to be fine-tuned beyond the max zoom, as IE does not support non-scaling strokes (SVG 1.2 Tiny).
The algorithm for calculating the edge width needed to be based on the mapping of one map px to one viewport pixel.

For now the extraZoom option is only enabled for WebKit browsers (chrome, safari), where it is working ok.

Also implemented an alternative for reverse scaling nodes by using the line-hack to create nodes with non-scaling-stroke vector effect
This works and this option is available for non-IE (SVG12T) browsers.
But, note that with this option the nodesGrow option won't be functional, the nodes will have thesame size on each zoom level.

Regardless of the option used to reverse scale nodes
----------------------------------------------------
Remaining bug in firefox due to the fact that at a given zoomlevel reverse scaled nodes won't receive mouse  events in FF

SEE also https://codepen.io/rbuhrs/pen/LjpPBb/

Note that this issue only occurs beyond zoomstep in var revScaleNodesJSStep which is also the tipping point for min PX size