Changes for extra zoom working on FF, option to use fixed scaled html overlays on nodes, prepare for animations.
Html overlay for injecting html into the graph (at node locations) from calling apps for use on click and hovers

Workaround for catching clicks on nodes in FF at large zoom levels, this is a FF bug when the map is scaled very large and as a consequence the nodes are scaled a lot in reverse.

Fixed node scaling in JS, getting the correct nodes in the viewport, so extra zoom can also be used in FF and IE now.

Preparing animated zoom / pan support.